requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.6
click==8.0.1
cloudinary==1.26.0
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-bootstrap-v5==1.0.5
django-localflavor==3.1
django-on-heroku==1.1.2
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.15
idna==3.2
importlib-metadata==3.10.1
lorem==0.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.3.0
pre-commit==2.15.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-dotenv==0.19.1
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
whitenoise==5.3.0
zipp==3.6.0
